Attempt at Flee
As of right now i may or may not be able to know the position of the circle center and I am planning on making it so that if the player is in the outside of the circle they will prioritize running back into it and if they see another character they will run a certain distance or preferably in a certain direction. I want to use flee for any time a contestant needs to run away.